% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.ridge.cv.R
\name{logreg.ridge.cv}
\alias{logreg.ridge.cv}
\title{logreg.ridge.cv}
\usage{
logreg.ridge.cv(y, X, lambdas, nfolds = 5, plot = F)
}
\arguments{
\item{y}{the input y as response}

\item{X}{the input X to fit model}

\item{lambdas}{a vector specifies the lambda candidates}

\item{nfolds}{an int sepecific how many folds you want to split the data}

\item{plot}{this boolean variable specifies if you want to plot the result}
}
\value{
the function will return a list of following elements: lambda.min, the best lambda in the candidates; err the error list corresponding to the lambda vector, and lambda.vec specifies the lambda vector as input
}
\description{
This function logreg.ridge.cv() works to perform cross-validation to select the best lambda
}
\examples{
x=matrix(rnorm(100),20,5)
y=sample(c(0,1),20,replace=TRUE)
lambdas=c(0.1,0.004,0.7,4)
logreg.ridge.cv(y,x,lambdas,plot=TRUE)
}
