% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.ridge.fit.R
\name{logreg.ridge.fit}
\alias{logreg.ridge.fit}
\title{logreg.ridge.fit}
\usage{
logreg.ridge.fit(
  y,
  X,
  lambda,
  epochs = 10,
  tol = 1e-05,
  beta0 = rep(0, dim(X)[2])
)
}
\arguments{
\item{y}{input y, this is the label of the class, it should be a vector of either zero or one}

\item{X}{input X, this is a matrix of training X.}

\item{lambda}{the coefficient of how severe the penalty should be}

\item{epochs}{this value specifies how much iteration will we do to update our coefficients}

\item{tol}{this input means tolerance, which determines early stop in training}

\item{beta0}{this is the initialization of beta estimators, sometimes user may want to further train a specific beta, and they can use it as input.}
}
\value{
it will return a list of estimated coefficients, the predicted score y.hat, the predicted class yhat and classification error.
}
\description{
This function takes y,x and lambda to perform a logistic regression w/wo l2 penalty
The input data should be a numeric matrix without factoer level, though we can use the transform method to encode it.
This function will return a list including coefficients of estimates, yhat, and predicted class, and classification error
}
\examples{
x=matrix(rnorm(100),20,5)
y=sample(c(0,1),20,replace=TRUE)
logreg.ridge.fit(y,x,lambda=0)
}
