% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.ridge.predict.R
\name{logreg.ridge.predict}
\alias{logreg.ridge.predict}
\title{logreg.ridge.predict
This function works as prediction and possible return classification error}
\usage{
logreg.ridge.predict(logreg_object, X.new, y.new = "None")
}
\arguments{
\item{logreg_object}{this is the object obtained from running fit.logreg.ridge() function}

\item{X.new}{this is the matrix data we want to compute.}

\item{y.new}{this is optional y. If a y.new is specified, then the function will use it to compute error.}
}
\value{
it will return a list object with three elements: y.hat the predicted score, pred the predicted class, err if y is pecified then this is the error, otherwise its 'None'
}
\description{
logreg.ridge.predict
This function works as prediction and possible return classification error
}
\examples{
x=matrix(rnorm(100),20,5)
y=sample(c(0,1),20,replace=TRUE)
logreg_object = logreg.ridge.fit(y,x,lambda=0)
X_new = x
logreg.ridge.predict( logreg_object, X_new, y )
}
